Remove legacy Table/Spreadsheet; rename DataSpreadsheet -> Spreadsheet
The virtualized DataGrid + its editable subclass fully cover the old
non-virtualized Table/Spreadsheet, so remove the duplicate set:

- delete lib/komps/table(.js) and lib/komps/spreadsheet(.js) with their
  sub-components and the Table-based plugins under table/plugins/
- rename DataSpreadsheet -> Spreadsheet (file, dir, class, tag
  komp-data-spreadsheet -> komp-spreadsheet), reusing the freed name;
  DataGrid keeps its name
- komps/plugins now re-exports the DataGrid plugins; add a komps/data-grid
  subpath export; drop komps/table
- update barrel, README, docs example, CLAUDE.md, and CHANGELOG
