+++ lib/komps.js
@@ -1,15 +1,13 @@
 export { default as AutoGrid } from './komps/auto-grid.js';
 export { default as Modal } from './komps/modal.js';
 export { default as Floater } from './komps/floater.js';
-export { default as Table } from './komps/table.js';
 export { default as Input } from './komps/input.js';
 export { default as Tooltip } from './komps/tooltip.js';
 export { default as Dropdown } from './komps/dropdown.js';
 export { default as NotificationCenter } from './komps/notification-center.js';
 export { default as Dropzone } from './komps/dropzone.js';
-export { default as Spreadsheet } from './komps/spreadsheet.js';
 export { default as DataGrid } from './komps/data-grid.js';
-export { default as DataSpreadsheet } from './komps/data-spreadsheet.js';
+export { default as Spreadsheet } from './komps/spreadsheet.js';
 export { default as ContentArea } from './komps/content-area.js';
 export { default as Element } from './komps/element.js';
 export { default as FormField } from './komps/form-field.js';
