+++ lib/komps/data-grid.js
@@ -19,7 +19,7 @@
  * @param {function} [options.rowClass] - `(record) => class` applied to each row element (string, token array, or {@link State}); re-evaluated each time a row mounts
  * @param {function} [options.empty] - `() => content` rendered in the body when there are no rows
  * @param {function} [options.loading] - `() => content` rendered in the body while `data` is still resolving (rows not yet built); removed once rows are in place
- * @param {function} [options.loadRecords] - `(records) => Promise` batch loader called with the records of rows entering the window (overscan included), before their cells render. Hydrate the records in place (e.g. fetch and merge nested associations) and resolve when done; cells render once the batch settles. Mount-driven only: records accessed while off-screen (e.g. DataSpreadsheet copy across an off-screen range) are not loaded.
+ * @param {function} [options.loadRecords] - `(records) => Promise` batch loader called with the records of rows entering the window (overscan included), before their cells render. Hydrate the records in place (e.g. fetch and merge nested associations) and resolve when done; cells render once the batch settles. Mount-driven only: records accessed while off-screen (e.g. Spreadsheet copy across an off-screen range) are not loaded.
  * @param {function} [options.loadingRow] - `(record, row) => content` rendered inside a row while `loadRecords` is resolving its record; replaced by cells once the load settles
  *
  * @fires DataGrid#rendered
