+++ lib/komps/data-grid/column.js
@@ -1,10 +1,9 @@
 /**
  * Column controller for a {@link DataGrid}.
  *
- * A standalone, persistent plain class (one per column) — mirrors the role of
- * `TableColumn`, but is **not** related to it (DataGrid is built standalone, with
- * no reuse of `table/`). It holds column config + behavior and tracks the cells
- * that are currently *mounted* for this column. Off-screen cells don't exist as
+ * A standalone, persistent plain class (one per column). It holds column config +
+ * behavior and tracks the cells that are currently *mounted* for this column.
+ * Off-screen cells don't exist as
  * DOM, so the full-dataset view is provided lazily via {@link DataGridColumn#cellHandles}.
  *
  * @class DataGridColumn
@@ -20,7 +19,7 @@
  * @param {string} [options.class] - extra class(es) applied to header + cells
  * @param {function} [options.unnest] - expand the record's nested collection into stacked sub-rows for this cell (SQL `UNNEST`/`explode`). A function `(record) => Array` returning the sub-records to stack (read-only display in v1)
  * @param {function} [options.splitInto] - **deprecated** alias for `unnest`
- * @param {function} [options.record] - resolve the record this column reads/writes from the row's record (e.g. a row `TIM` → its `Trait`). A function `(rowRecord) => record | Promise<record>`. When omitted the column operates on the row's record directly. `renderContent`/`valueFor` and (for {@link DataSpreadsheet}) `input`/`paste`/`clear`/`toggle` all operate on the resolved record.
+ * @param {function} [options.record] - resolve the record this column reads/writes from the row's record (e.g. a row `TIM` → its `Trait`). A function `(rowRecord) => record | Promise<record>`. When omitted the column operates on the row's record directly. `renderContent`/`valueFor` and (for {@link Spreadsheet}) `input`/`paste`/`clear`/`toggle` all operate on the resolved record.
  */
 
 import { isFunction } from '../../support.js'
