+++ lib/komps/data-grid/plugins/collapsible.js
@@ -1,5 +1,5 @@
 /**
- * A plugin to make a {@link DataGrid}'s (and its subclasses', e.g. {@link DataSpreadsheet})
+ * A plugin to make a {@link DataGrid}'s (and its subclasses', e.g. {@link Spreadsheet})
  * rows **collapsible** — the windowed counterpart of the Table `collapsible` plugin.
  *
  * The app sets a height budget for rows via `collapseTo` (any valid CSS size). Rows whose
@@ -38,7 +38,7 @@ import { createElement, listenerElement } from 'dolla'
 
 export default function (proto) {
     // assignableAttributes is shared up the prototype chain; clone before adding so a
-    // plugin included on a subclass (DataSpreadsheet) doesn't leak options onto DataGrid.
+    // plugin included on a subclass (Spreadsheet) doesn't leak options onto DataGrid.
     if (!Object.hasOwn(this, 'assignableAttributes')) {
         this.assignableAttributes = { ...this.assignableAttributes }
     }
@@ -191,7 +191,7 @@ export default function (proto) {
             })
         })
         if (cell.classList.contains('frozen')) toggle.style.left = cell.style.left
-        // Keep the toggle from starting a selection / edit underneath (DataSpreadsheet).
+        // Keep the toggle from starting a selection / edit underneath (Spreadsheet).
         toggle.addEventListener('pointerdown', e => e.stopPropagation())
         toggle.addEventListener('mousedown', e => e.stopPropagation())
         row.element.prepend(toggle)
@@ -219,7 +219,7 @@ export default function (proto) {
             justify-content: end;
             text-align: center;
             grid-row: 1 / -1;
-            /* Above DataSpreadsheet's active cell (z 3) so the button stays clickable
+            /* Above Spreadsheet's active cell (z 3) so the button stays clickable
                when its cell is selected, but below frozen cells (z 5) so a scrolling
                toggle still slides under frozen columns. */
             z-index: 4;
