+++ lib/komps/data-grid/plugins/reorderable.js
@@ -1,5 +1,5 @@
 /**
- * A plugin to make a {@link DataGrid} (and its subclasses, e.g. {@link DataSpreadsheet})
+ * A plugin to make a {@link DataGrid} (and its subclasses, e.g. {@link Spreadsheet})
  * **reorderable** along columns and/or rows.
  *
  * @function Plugin/DataGridReorderable
@@ -12,10 +12,10 @@
  * @fires Plugin/DataGridReorderable#rowReordered
  *
  * @example <caption>JS</caption>
- * import DataSpreadsheet from 'komps/komps/data-spreadsheet.js'
+ * import Spreadsheet from 'komps/komps/spreadsheet.js'
  * import { reorderable } from 'komps/komps/data-grid/plugins.js'
- * DataSpreadsheet.include(reorderable)
- * new DataSpreadsheet({
+ * Spreadsheet.include(reorderable)
+ * new Spreadsheet({
  *     reorder: ['columns'],    // columns only; omit for both
  *     data: [...],
  *     columns: [
@@ -48,7 +48,7 @@ const clamp = (v, lo, hi) => Math.max(lo, Math.min(v, hi))
 
 export default function (proto) {
     // assignableAttributes is shared up the prototype chain; clone before adding so a
-    // plugin included on a subclass (DataSpreadsheet) doesn't leak `reorder` onto DataGrid.
+    // plugin included on a subclass (Spreadsheet) doesn't leak `reorder` onto DataGrid.
     if (!Object.hasOwn(this, 'assignableAttributes')) {
         this.assignableAttributes = { ...this.assignableAttributes }
     }
