+++ lib/komps/data-grid/plugins/resizable.js
@@ -1,5 +1,5 @@
 /**
- * A plugin to make a {@link DataGrid} (and its subclasses, e.g. {@link DataSpreadsheet})
+ * A plugin to make a {@link DataGrid} (and its subclasses, e.g. {@link Spreadsheet})
  * **resizable** along columns and/or rows.
  *
  * @function Plugin/DataGridResizable
@@ -13,10 +13,10 @@
  * @fires Plugin/DataGridResizable#rowResized
  *
  * @example <caption>JS</caption>
- * import DataSpreadsheet from 'komps/komps/data-spreadsheet.js'
+ * import Spreadsheet from 'komps/komps/spreadsheet.js'
  * import { resizable } from 'komps/komps/data-grid/plugins.js'
- * DataSpreadsheet.include(resizable)
- * new DataSpreadsheet({
+ * Spreadsheet.include(resizable)
+ * new Spreadsheet({
  *     resize: ['columns'],     // columns only; omit for both
  *     data: [...],
  *     columns: [
@@ -46,7 +46,7 @@ import { createElement } from 'dolla'
 
 export default function (proto) {
     // assignableAttributes is shared up the prototype chain; clone before adding so a
-    // plugin included on a subclass (DataSpreadsheet) doesn't leak `resize` onto DataGrid.
+    // plugin included on a subclass (Spreadsheet) doesn't leak `resize` onto DataGrid.
     if (!Object.hasOwn(this, 'assignableAttributes')) {
         this.assignableAttributes = { ...this.assignableAttributes }
     }
@@ -256,7 +256,7 @@ export default function (proto) {
         this.applyColumnGeometry()   // re-publish --dg-template-columns / --dg-width (header + rows inherit)
         this.refreshFrozenColumns()  // sticky lefts of later frozen columns shift with the new width
         this.sizeBody()
-        this.paintSelection?.()      // DataSpreadsheet: keep the range box aligned to the new geometry
+        this.paintSelection?.()      // Spreadsheet: keep the range box aligned to the new geometry
         this.trigger('columnResized', { detail: { column, width } })
     }
