+++ lib/komps/data-grid/row.js
@@ -2,8 +2,8 @@
  * Row controller for a {@link DataGrid}.
  *
  * A standalone, persistent plain class — one per **data row** — mirroring the
- * {@link DataGridColumn} controller pattern (and deliberately **not** `TableRow`,
- * which is itself an element). It is the source-of-truth object for a row and
+ * {@link DataGridColumn} controller pattern (a controller object, not itself an
+ * element). It is the source-of-truth object for a row and
  * serves as the row *handle*: code may hold it whether or not the row is on-screen.
  *
  * It **owns its measured `height`** (`null` until measured — mirroring how a column
