+++ lib/komps/plugins.js
@@ -1,3 +1,1 @@
-export { default as resizable } from './table/plugins/resizable.js';
-export { default as reorderable } from './table/plugins/reorderable.js';
-export { default as collapsible } from './table/plugins/collapsible.js';
+export { resizable, reorderable, collapsible } from './data-grid/plugins.js';
