+++ lib/komps/spreadsheet.js
@@ -1,139 +1,165 @@
 /**
- * A {@link Table} with editable cells
+ * An editable {@link DataGrid} — the virtualized counterpart to {@link Spreadsheet}.
  *
  * @class Spreadsheet
- * @extends Table
+ * @extends DataGrid
  *
  * @param {Object} [options={}]
- * @param {boolean} [options.scrollSnap=false] - Enable scroll snapping (WIP, needs to update scroll-start when frozen rows/columns stick)
- * @param {Array} [options.columns=[]] - Array of objects that initialize {@link SpreadsheetColumn Columns} that manage the rendering of {@link SpreadsheetCell Cells}
- * @param {function} [options.readonly] - Function called with each `record` to determine whether its row should render readonly cells. Forwarded to each {@link TableRow} and resolved per-record in `row.renderCell` before delegating to `column.renderCell`. Per-cell `readonly` overrides (e.g. {@link ReadonlyColumn}, split-cell placeholders) still win.
+ * @param {function} [options.readonly] - Function called with each `record` to determine whether its row should render readonly (non-editable) cells. Resolved per-record before editing, pasting, or clearing. A column that is explicitly non-editable (`editable: false`) stays non-editable regardless; per-record `readonly` additionally makes otherwise-editable cells readonly.
  *
- * @example <caption>JS</caption>
- * new Spreadsheet({
- *     data: [{
- *         name: "Corey Seager",
- *         team: "Texas Rangers",
- *         ops: 1.023
- *     }],
- *     columns: [{
- *         attribute: 'name',
- *         header: 'Name'
- *     }, {
- *         attribute: 'team',
- *         header: 'Team',
- *     }]
- * })
- *
- * @todo collapse selectedCells when pasting to match dimensions of copy (like google sheets)
- * @todo expand selectedCells when pasting to match dimensions of copy (like google sheets)
- * @todo select multiple columns reorder, only moves one
- * @todo edit header needs to fire event, or provide object to change
- * @todo Tab to a cell causes sheet to scroll instead of using single cell scroll of focusAdjacentCell
- * @todo insert/remove column
- * @todo copy/paste split row
- * @todo mouse select cells, then arrow select, arrow selection starts at original start of mouse and not end
- * @todo selection outline overlays and doesn't track with frozen cells
- * @todo implement ClipboardAPI / ClipboardItem for copy/paste
+ * @fires Spreadsheet#cellChanged
+ * @fires Spreadsheet#invalidPaste
  */
 
-import { createElement, listenerElement, content, insertAfter } from 'dolla';
-import { result, except, placeCaretAtEnd, closest, groupBy, isInView, splitByUnquotedChar, uniq } from '../support.js';
-import Table from './table.js';
-import Floater from './floater.js';
-import Modal from './modal.js';
-import Cell from './spreadsheet/cell.js';
-import Column from './spreadsheet/column.js';
-import ReadonlyColumn from './spreadsheet/columns/readonly-column.js';
-import NumberColumn from './spreadsheet/columns/number-column.js';
-import CheckboxColumn from './spreadsheet/columns/checkbox-column.js';
-import SelectColumn from './spreadsheet/columns/select-column.js';
-import { reorderable, resizable } from './plugins.js';
+import { createElement, listenerElement } from 'dolla'
 
-export default class Spreadsheet extends Table {
+import { splitByUnquotedChar } from '../support.js';
+
+import DataGrid from './data-grid.js'
+import Floater from './floater.js'
+import Modal from './modal.js'
+import SpreadsheetCell from './spreadsheet/cell.js'
+
+import SpreadsheetColumn from './spreadsheet/column.js'
+import NumberColumn from './spreadsheet/columns/number-column.js'
+import SelectColumn from './spreadsheet/columns/select-column.js'
+import CheckboxColumn from './spreadsheet/columns/checkbox-column.js'
+import ReadonlyColumn from './spreadsheet/columns/readonly-column.js'
+
+export default class Spreadsheet extends DataGrid {
     static tagName = 'komp-spreadsheet'
 
-    static assignableAttributes = {
-        scrollSnap: { type: 'boolean', default: false, null: false },
-        readonly: { type: 'function', default: null, null: true }
-    }
-    
+    static Cell = SpreadsheetCell
+    static Column = SpreadsheetColumn
     static columnTypeRegistry = {
-        select: SelectColumn,
+        default: SpreadsheetColumn,
         number: NumberColumn,
+        select: SelectColumn,
         checkbox: CheckboxColumn,
         radio: CheckboxColumn,
-        readonly: ReadonlyColumn,
-        default: Column
+        readonly: ReadonlyColumn
     }
+    static cellSelector = this.Cell.tagName
+
+    static assignableAttributes = {
+        readonly: { type: 'function', default: null, null: true }
+    }
+
+    static events = ['invalidPaste', 'cellChanged']
+
+    constructor(...args) {
+        super(...args)
+        this.selection = null        // { anchor:{row,col}, focus:{row,col} } | null
+        this.active = null           // { row, col } — the active cell: focused + outlined, the selection's anchor
+        this.cursor = null           // { row, col } — the moving edge of the range while extending (drag / shift+arrow)
+        this._editing = null         // { cell, column, record, input } | null
+        this._copying = false        // copy pending → the selection indicator shows the dashed marquee
+    }
+
+    async initialize(...args) {
+        const result = await super.initialize(...args)
+        this.setupSelection()
+        return result
+    }    
     
-    static events = ["invalidPaste"]
+    /* -----------------------------------------------------------------
+       Event wiring
+    ----------------------------------------------------------------- */
+    setupSelection() {
+        const sel = this.constructor.cellSelector
     
-    initialize () {
-        if (this.scrollSnap) {
-            this.classList.add('scroll-snap')
-        }
-        this.addEventListenerFor(this.cellSelector, 'mousedown', e => {
-            if (e.button == 0) {
-                this.clearSelectedCells()
-                this.activateMouseCellSelection(e.delegateTarget, e)
+        this.addEventListenerFor(sel, 'mousedown', e => {
+            if (e.button !== 0) return
+            const cell = e.delegateTarget
+            const pos = { row: cell.rowIndex, col: cell.cellIndex }
+            if (e.shiftKey && this.active) {
+                // Shift-click extends the range from the active cell, which stays.
+                this.cursor = pos
+                this.setSelection(this.active, pos)
+                cell.focus({ preventScroll: true })
+            } else {
+                // Plain click collapses the selection onto the clicked cell.
+                this.focusCellAt(pos.col, pos.row)
             }
-        })
-        this.addEventListenerFor(this.cellSelector, 'dblclick', e => {
-            e.delegateTarget.activate();
-        })
-        
-        this.addEventListenerFor(`${this.tagName}-reorder-handle`, ['click', 'contextmenu'], e => {
-            this.activateContextMenu(e)
-            e.preventDefault()
-        })
-        
-        this.addEventListenerFor(this.cellSelector, 'contextmenu', e => {
-            this.activateContextMenu(e)
-            if (!e.delegateTarget.classList.contains('selected')) {
-                this.clearSelectedCells()
-                this.selectCells(e.delegateTarget, e.delegateTarget, true)
-            }
-            e.preventDefault()
-        })
-        this.addEventListenerFor(this.cellSelector, 'keydown', async e => {
-            if (["ArrowRight", "ArrowLeft", "ArrowDown", "ArrowUp"].includes(e.key)) {
-                e.preventDefault()
-                const targetCell = e.delegateTarget.focusAdjacentCell(e.key.replace('Arrow', '').toLowerCase())
-                if (e.shiftKey) {
-                    this.selectStartCell = this.selectStartCell || e.delegateTarget
-                    this.selectCells(this.selectStartCell, targetCell, true)
-                } else {
-                    this.clearSelectedCells()
+    
+            // Drag moves the cursor (range edge) from the active cell, which stays put.
+            const move = ev => {
+                const c = ev.target instanceof Element ? ev.target.closest(sel) : null
+                if (c && c.column) {
+                    this.cursor = { row: c.rowIndex, col: c.cellIndex }
+                    this.setSelection(this.active, this.cursor)
                 }
-            } else if (e.key == "Enter" && e.delegateTarget == this.getRootNode().activeElement) {
-                e.preventDefault();
-                e.delegateTarget.activate();
-                this._enterDown = true;
-                this.addEventListener('keyup', e => {
-                    delete this._enterDown;
-                }, { once: true })
-            } else if (e.key == "Tab") {
-                e.preventDefault()
-                e.delegateTarget.focusAdjacentCell(e.shiftKey ? 'left' : 'right')
-            } else if (e.key == "Escape") {
-                this.clearSelectedCells('')
-            } else if (['Backspace', 'Delete', 'Clear'].some(x => x == e.key)) {
-                e.delegateTarget.clear()
-            } else if (e.key.length == 1 && [e.metaKey,e.ctrlKey,e.altKey].every(x => x == false)) {
-                e.delegateTarget.activate({
-                    value: e.key
-                });
-                e.preventDefault()
+            }
+            const up = () => this.removeEventListener('mousemove', move)
+            this.addEventListener('mousemove', move)
+            this.getRootNode().addEventListener('mouseup', up, { once: true })
+        })
+    
+        this.addEventListenerFor(sel, 'dblclick', e => this.activateCell(e.delegateTarget))
+    
+        // Only seed `active` on a genuine first focus (e.g. tabbing in). Don't override
+        // it when focus follows the moving edge during a range extend.
+        this.addEventListenerFor(sel, 'focusin', e => {
+            if (this.active == null) {
+                const cell = e.delegateTarget
+                this.active = { row: cell.rowIndex, col: cell.cellIndex }
             }
         })
-        return super.initialize()
-    }
     
-    connected (...args) {
-        if (!this.getRootNode()) { return }
-        this.cleanupEventListenerFor(this.getRootNode(), 'paste', e => {
-            if (this.contains(document.activeElement) && document.activeElement instanceof Cell) {
+        this.addEventListenerFor(sel, 'keydown', e => {
+            if (this._editing) return
+            const cell = e.delegateTarget
+            const key = e.key
+            if (['ArrowUp', 'ArrowDown', 'ArrowLeft', 'ArrowRight'].includes(key)) {
+                e.preventDefault()
+                this.moveActive(key.replace('Arrow', '').toLowerCase(), e.shiftKey)
+            } else if (key === 'Enter') {
+                e.preventDefault()
+                this.activateCell(cell)
+                // The same Enter press fires a keyup on the editor we just focused; swallow it
+                // (once, in capture so it beats the editor's own listener) so it can't drive
+                // editor behavior — e.g. a SearchField selecting its first result on keyup and
+                // instantly closing the cell we just opened.
+                this.getRootNode().addEventListener('keyup', ev => {
+                    if (ev.key === 'Enter') { ev.stopImmediatePropagation(); ev.preventDefault() }
+                }, { once: true, capture: true })
+            } else if (key === 'Tab') {
+                e.preventDefault()
+                this.moveActive(e.shiftKey ? 'left' : 'right')
+            } else if (key === 'Escape') {
+                // First Escape cancels the copy marquee; otherwise clears the selection.
+                if (this._copying) { this._copying = false; this.paintSelection() }
+                else this.clearSelection()
+            } else if (key === 'Backspace' || key === 'Delete') {
+                e.preventDefault()
+                this.clearSelectedCells()
+            } else if (key.length === 1 && !e.metaKey && !e.ctrlKey && !e.altKey) {
+                e.preventDefault()
+                this.activateCell(cell, { value: key })
+            }
+        })
+    
+        this.addEventListenerFor(sel, 'contextmenu', e => {
+            e.preventDefault()
+            const cell = e.delegateTarget
+            const pos = { row: cell.rowIndex, col: cell.cellIndex }
+            // If the right-clicked cell isn't part of the current selection, collapse to it.
+            const r = this.range()
+            const inSel = !!r && pos.row >= r.r0 && pos.row <= r.r1 && pos.col >= r.c0 && pos.col <= r.c1
+            if (!inSel) {
+                this.focusCellAt(pos.col, pos.row)
+            }
+            this.activateContextMenu(e)
+        })
+
+        this.addEventListener('copy', e => {
+            if (this.selection && this.contains(this.getRootNode().activeElement)) {
+                e.preventDefault();
+                this.copyCells()
+            }
+        })
+        this.addEventListener('paste', e => {
+            if (this.selection && !this._editing && this.contains(this.getRootNode().activeElement)) {
                 e.preventDefault()
                 if (e.clipboardData.files.length > 0) {
                     this.pasteData(e.clipboardData.files)
@@ -142,196 +168,472 @@ export default class Spreadsheet extends Table {
                 }
             }
         })
-        
-        this.cleanupEventListenerFor(this.getRootNode(), 'copy', e => {
-            if (this.contains(document.activeElement) && document.activeElement instanceof Cell) {
-                e.preventDefault()
-                this.copyCells()
-            }
-        })
-        return super.connected(...args)
     }
 
-    async renderRow (record, index) {
-        return super.renderRow(record, index, { readonly: this.readonly })
+    /** Make cells focusable, set once per physical element (pooled cells keep it). */
+    acquireCell() {
+        const cell = super.acquireCell()
+        if (cell.tabIndex < 0) cell.tabIndex = 0 // freshly created cells default to -1
+        return cell
     }
 
-    /* -------------------
-        Cell Selection
-    ------------------- */
-    selectCells (start, end, isDone=false) {
-        if (end == undefined) end = start;
-        
-        const isHeader = start.localName.includes('header') || end.localName.includes('header')
-        const selectClass = isDone ? 'selected' : 'selecting';
+    /* -----------------------------------------------------------------
+       Window hook — repaint selection whenever the mounted set changes
+    ----------------------------------------------------------------- */
+    updateWindow() {
+        super.updateWindow()
+        this.paintSelection()
+    }
 
-        this.queryCells('.' + selectClass).forEach(el => el.classList.remove(selectClass))
-        this.querySelectorAll(`${this.localName}-outline:not(.copy)`).forEach(el => el.remove())
-        if (start.localName.includes('header')) {
-            end = this.at(end.cellIndex, this.rows.length)
+    /* -----------------------------------------------------------------
+       Selection range (over the model) + paint onto mounted cells
+    ----------------------------------------------------------------- */
+    range(sel = this.selection) {
+        if (!sel) return null
+        return {
+            r0: Math.min(sel.anchor.row, sel.focus.row), r1: Math.max(sel.anchor.row, sel.focus.row),
+            c0: Math.min(sel.anchor.col, sel.focus.col), c1: Math.max(sel.anchor.col, sel.focus.col)
         }
-        if (end.localName.includes('header')) {
-            start = this.at(start.cellIndex, this.rows.length)
+    }
+
+    setSelection(anchor, focus = anchor) {
+        this._copying = false // any new/changed selection cancels the copy marquee
+        this.selection = { anchor, focus }
+        this.paintSelection()
+    }
+    clearSelection() {
+        this._copying = false
+        this.selection = null
+        this.paintSelection()
+    }
+
+    /** Whether a record's row renders readonly (non-editable) cells. */
+    isRecordReadonly(record) {
+        return this.readonly ? !!this.readonly(record) : false
+    }
+
+    paintSelection() {
+        if (!this.body) return
+        const r = this.range()
+        const a = this.active
+        for (const row of this.mountedRows) {
+            for (const column of this.columns) {
+                const cell = row.cellOf(column)
+                if (!cell) continue
+                const inSel = !!r && row.index >= r.r0 && row.index <= r.r1 && column.index >= r.c0 && column.index <= r.c1
+                cell.classList.toggle('selected', inSel)
+                cell.classList.toggle('active', !!a && a.row === row.index && a.col === column.index)
+                // `readonly` is a per-binding property, applied in SpreadsheetCell#bind.
+            }
         }
-        
-        const cells = this.slice(start, end)
-        cells.forEach(cell => {
-            cell.classList.add(selectClass)
+        this.renderSelectionIndicator(r)
+    }
+
+    /**
+     * The border box around the selection range — solid normally, dashed (`.copy`)
+     * while a copy is pending (the marquee).
+     */
+    renderSelectionIndicator(r = this.range()) {
+        if (!r) {
+            if (this._selectIndicator) { this._selectIndicator.remove(); this._selectIndicator = null }
+            return
+        }
+        const el = this._selectIndicator || (this._selectIndicator = this.body.appendChild(createElement(`${this.localName}-select-indicator`)))
+        const cg = this.columnGeometry, rg = this.rowGeometry
+        el.style.left = cg.offsetAt(r.c0) + 'px'
+        el.style.top = rg.offsetAt(r.r0) + 'px'
+        el.style.width = (cg.offsetAt(r.c1) + cg.sizeAt(r.c1) - cg.offsetAt(r.c0)) + 'px'
+        el.style.height = (rg.offsetAt(r.r1) + rg.sizeAt(r.r1) - rg.offsetAt(r.r0)) + 'px'
+        el.classList.toggle('copy', !!this._copying)
+    }
+
+    cellElementAt(row, col) {
+        const r = this.rows[row], c = this.columns[col]
+        return (r && c) ? r.cellOf(c) : null
+    }
+
+    /* -----------------------------------------------------------------
+       Navigation — move the active cell, scrolling off-screen into view
+    ----------------------------------------------------------------- */
+    moveActive(dir, extend = false) {
+        if (extend) {
+            // Extend: move the cursor (range edge); the active cell stays put.
+            const from = this.cursor || this.active
+            if (!from) return
+            const { row, col } = this.stepFrom(from, dir)
+            this.cursor = { row, col }
+            this.setSelection(this.active || this.cursor, this.cursor)
+            // Reveal the cursor, but keep focus + outline on the active cell. Fall back to
+            // the cursor only when the active has scrolled out of view — so we never show
+            // two outlines at once.
+            this.scrollCellIntoView(row, col)
+            this.updateWindow()
+            const activeCell = this.active && this.cellElementAt(this.active.row, this.active.col)
+            ;(activeCell || this.cellElementAt(row, col))?.focus({ preventScroll: true })
+        } else {
+            // Move the active cell; collapse the selection to it.
+            const from = this.active || (this.selection && this.selection.focus)
+            if (!from) return
+            const { row, col } = this.stepFrom(from, dir)
+            this.focusCellAt(col, row)
+        }
+    }
+
+    /** One step from {row,col} in a direction, clamped to the grid. */
+    stepFrom(from, dir) {
+        let row = from.row, col = from.col
+        if (dir === 'up') row--
+        else if (dir === 'down') row++
+        else if (dir === 'left') col--
+        else if (dir === 'right') col++
+        return {
+            row: Math.max(0, Math.min(row, this.rows.length - 1)),
+            col: Math.max(0, Math.min(col, this.columns.length - 1))
+        }
+    }
+
+    /**
+     * Make (col, row) the active cell: collapse the selection onto it, scroll it
+     * into view, and move focus to it — the programmatic equivalent of a plain
+     * left-click, and the shared path behind a click, an un-extended arrow/Tab
+     * move, and a right-click collapse. Argument order matches {@link DataGrid#at}.
+     *
+     * @param {number} col - Column index (0-based)
+     * @param {number} row - Row index (0-based)
+     * @returns {SpreadsheetCell|null} the focused cell, if currently mounted
+     */
+    focusCellAt(col, row) {
+        this.active = { row, col }
+        this.cursor = { row, col }
+        this.setSelection({ row, col }) // collapse the range; clears the copy marquee
+        this.scrollCellIntoView(row, col)
+        let cell = this.cellElementAt(row, col)
+        // Single-step moves land within overscan, so the cell is already mounted.
+        // Only force a synchronous window pass for a target beyond overscan (a far
+        // jump, or overscan: 0); otherwise the scroll event re-settles the window.
+        if (!cell) {
+            this.updateWindow()
+            cell = this.cellElementAt(row, col)
+        }
+        if (cell) cell.focus({ preventScroll: true })
+        return cell
+    }
+
+    frozenWidth() {
+        let w = 0
+        for (const c of this.columns) if (c.frozen) w += this.columnGeometry.sizeAt(c.index)
+        return w
+    }
+
+    scrollCellIntoView(row, col) {
+        const headerH = this.body.offsetTop
+        const top = headerH + this.rowGeometry.offsetAt(row)
+        const bottom = top + this.rowGeometry.sizeAt(row)
+        if (top < this.scrollTop + headerH) this.scrollTop = top - headerH
+        else if (bottom > this.scrollTop + this.clientHeight) this.scrollTop = bottom - this.clientHeight
+
+        if (!(this.columns[col] && this.columns[col].frozen)) {
+            const frozenW = this.frozenWidth()
+            const left = this.columnGeometry.offsetAt(col)
+            const right = left + this.columnGeometry.sizeAt(col)
+            if (left < this.scrollLeft + frozenW) this.scrollLeft = left - frozenW
+            else if (right > this.scrollLeft + this.clientWidth) this.scrollLeft = right - this.clientWidth
+        }
+    }
+
+    /* -----------------------------------------------------------------
+       Editing
+    ----------------------------------------------------------------- */
+    /**
+     * Begin editing a cell.
+     *
+     * The column's `input(record, session)` receives an editing **session** a custom
+     * editor can use to control completion (it also carries `cell`, `column`, `record`,
+     * `floater`, and `value` — the typed character seeding a type-to-edit):
+     *  - `session.commit({ move })` — end the edit, re-render the cell, optionally move the active cell
+     *  - `session.cancel()` — end the edit without applying the bound input's change
+     *  - `session.hold(promise)` — defer the closing re-render until the promise settles
+     *    (e.g. an async save), however the edit ends (commit, blur, Enter/Tab)
+     *
+     * Keyboard: Enter (without Shift) and Tab commit and navigate. An editor that wants a
+     * key for itself (e.g. Enter inserting a newline in a textarea) stops its propagation —
+     * or calls `preventDefault()` — on `keydown` and the spreadsheet leaves it alone.
+     *
+     * Editors may also dispatch a bubbling `commit` CustomEvent instead of using the
+     * session: `detail.move` and `detail.promise` map to `commit({ move })` and `hold()`.
+     */
+    async activateCell(cell, session = {}) {
+        if (this._editing) this._editing.floater.hide()
+        const { column } = cell
+        if (!column || cell.record == null || column.editable === false) return
+        if (this.isRecordReadonly(cell.record)) return
+
+        // `readonly` is gated on the cell's row record above; input/toggle/paste act on the
+        // column's (possibly resolved) record. Resolved + memoized via the row.
+        const record = await cell.row.resolvedRecordFor(column)
+        if (record == null) return
+
+        if (column.toggleable) {
+            column.toggle(record)
+            cell.render()
+            this.trigger('cellChanged', { detail: { cell, column, record } })
+            return
+        }
+
+        // The editing session handed to the column's input (see the method JSDoc). Its
+        // callbacks close over `floater`, which is assigned below — they can only run
+        // once the editor exists, so the forward reference is safe.
+        let floater
+        Object.assign(session, {
+            cell, column, record,
+            pending: null,
+            commit: opts => this.deactivateCell(opts), // arrow: keeps `this`
+            hold: promise => { session.pending = promise; return promise },
+            cancel: () => {
+                floater.setAttribute('preventChange', true) // bound inputs check this before writing
+                this.deactivateCell()
+            }
         })
-        
-        if (isDone) {
-            this.outlineCells(cells)
-        }
-    }
-    selectedCells () {
-        let cells = this.queryCells('.selected')
-        if (cells.length == 0) {
-            cells = this.queryCells(':focus')
-        }
-        return cells
-    }
-    clearSelectedCells (scope=':not(.copy)') {
-        delete this.selectStartCell
-        this.queryCells(`.selecting${scope}`).forEach(el => el.classList.remove('selecting'))
-        this.queryCells(`.selected${scope}`).forEach(el => el.classList.remove('selected'))
-        this.clearOutlines(scope)
-    }
-    activateMouseCellSelection (startingCell, e) {
-        const is_header = startingCell.localName.includes('header')
-        
-        const mouseOver = e => {
-            const currentCell = closest(e.target, this.cellSelector)
-            if (currentCell) {
-                this.selectCells(startingCell, currentCell)
-            }
-        }
-        
-        const mouseUp = e => {
-            const endCell = closest(e.target, this.cellSelector)
-            if (endCell != startingCell || endCell.localName.includes('header')) {
-                const selectingCells = this.queryCells('.selecting')
-                selectingCells.forEach(el => {
-                    el.classList.remove('selecting')
-                    el.classList.add('selected')
-                })
-                this.outlineCells(selectingCells)
-            }
-            this.removeEventListener('mouseover', mouseOver)
-            startingCell.focus()
-        }
-        
-        if (e.shiftKey) {
-            const currentCell = startingCell
-            startingCell = this.queryCell(':focus')
-            startingCell = startingCell || this.queryCell('.selecting')
-            e.preventDefault()
-            mouseOver({ target: currentCell })
-        } else if (is_header) {
-            mouseOver({target: startingCell})
-        }
-        
-        this.getRootNode().addEventListener('mouseup', mouseUp, {once: true})
-        this.addEventListener('mouseover', mouseOver)
-    }
-    clearOutlines (scope=':not(.copy)') {
-        this.querySelectorAll(`${this.localName}-outline${scope}`).forEach(el => el.remove())
-    }
-    outlineCells (cells) {
-        if (!Array.isArray(cells)) { cells = Array.from(cells) }
-        const groupNames = uniq(cells.map(c => c.parentElement.dataset.name))
-        
-        const offsetLeft = Math.min(...cells.map(c => c.offsetLeft))
-        const offsetTop = Math.min(...cells.map(c => c.offsetTop))
-        const offsetRight = Math.max(...cells.map(c => c.offsetLeft + c.offsetWidth))
-        const offsetBottom = Math.max(...cells.map(c => c.offsetTop + c.offsetHeight))
-        const outline = createElement(`${this.localName}-outline`, {
-            style: {
-                left: offsetLeft + "px",
-                top: offsetTop + "px",
-                width: offsetRight - offsetLeft + "px",
-                height: offsetBottom - offsetTop + "px",
-            }
-        })
-        this.append(outline)
-        
-        // There's a reason this was here, but not sure why, removed because it was causing weirdness with selecting groupIndex with introduction of collapsible - BE
-        // const cellsTop = Math.min(...cells.map(c => c.offsetTop))
-        // const cellsBottom = Math.max(...cells.map(c => c.offsetTop + c.offsetHeight))
-        // if (outline.offsetTop < cellsTop) {
-        //     outline.style.marginTop = cellsTop - outline.offsetTop + "px"
-        // }
-        // if (outline.offsetTop + outline.offsetHeight > cellsBottom) {
-        //     outline.style.marginBottom = ((outline.offsetTop + outline.offsetHeight) - cellsBottom) + "px"
-        // }
-        
-        return outline
-    }
-    
-    /* ---------------------
-        Context Menu
-    --------------------- */
-    activateContextMenu (e) {
-        const cell = e.delegateTarget.cell || e.delegateTarget
-        const menu = this.renderContextMenu(cell)
-        const targetBB = e.target.getBoundingClientRect()
-        if (this.contextMenu) {
-            this.contextMenu.hide()
-            delete this.contextMenu
-        }
-        this.contextMenu = new Floater({
-            content: menu,
+
+        // The session is the input's options bag too — `value` seeds type-to-edit. Columns
+        // must pick what they need off it (not spread it into Input/element options): its
+        // members would otherwise land as HTML attributes on the editor.
+        const input = column.input(record, session)
+        if (!input) return
+
+        floater = new Floater({
+            class: `${this.localName}-input`,
             anchor: cell,
-            offset: {
-                mainAxis: e.offsetX - targetBB.width,
-                crossAxis: e.offsetY
-            },
-            placement: 'right-start',
-            shift: false,
-            flip: true,
-            autoPlacement: false,
+            container: this.utilities, // grid-level layer above the header/frozen z-stack
+            content: { tag: 'label', content: input, class: column.class },
+            placement: 'bottom-start',
+            autoPlacement: { alignment: 'start', allowedPlacements: ['top', 'bottom'] },
             removeOnBlur: true,
-            autoUpdate: false
+            offset: ({ rects }) => ({ mainAxis: -rects.reference.height }),
+            style: {
+                // Resolved px from geometry (numbers), not the raw config: column.width
+                // may be a CSS px string ("200px" + "px" → "200pxpx") and row.height is
+                // null until measured — either would yield an invalid length and drop the rule.
+                '--cell-width': this.columnGeometry.sizeAt(column.index) + "px",
+                '--cell-height': this.rowGeometry.sizeAt(cell.rowIndex) + "px"
+            },
+            onHidden: async e => {
+                cell.tabIndex = 0
+                if (this._editing && this._editing.floater === floater) this._editing = null
+                // Hold the re-render for work the editor registered via session.hold (e.g.
+                // an async save) so the cell reflects the settled value — this also covers
+                // the blur-close path, which never goes through deactivateCell.
+                if (session.pending) { try { await session.pending } catch { /* render whatever state remains */ } }
+                cell.render() // reflect the value the bound input wrote (or the unchanged value on cancel)
+                this.trigger('cellChanged', { detail: { cell, column, record } })
+            }
         })
-        this.append(this.contextMenu)
+        this._editing = { cell, column, record, floater, session }
+        cell.tabIndex = -1
+        
+        floater.addEventListener('keydown', e => {
+            e.stopPropagation()
+            if (e.defaultPrevented) return
+            if (e.key === 'Enter' && !e.shiftKey) {
+                requestAnimationFrame(() => {
+                    e.preventDefault();
+                    session.commit({move: 'down'})
+                })
+            }
+            else if (e.key === 'Tab') {
+                e.preventDefault();
+                session.commit({move: e.shiftKey ? 'left' : 'right'})
+            }
+            else if (e.key === 'Escape') {
+                e.preventDefault();
+                session.cancel()
+            }
+        })
+
+        // Event-driven sugar over the session API: an editor may dispatch a bubbling
+        // `commit` CustomEvent when its (possibly async) work is done.
+        floater.addEventListener('commit', e => {
+            if (e.detail?.promise) session.hold(e.detail.promise)
+            this.deactivateCell({ move: e.detail?.move })
+        })
+
+        floater.show() // appends into this.utilities (see container option above)
+
+        // Focus the editor once it's positioned. A column may return a single bound element
+        // or several nodes (e.g. an input plus an attribution badge/note), so query the
+        // rendered editor: prefer the element the column marked `[autofocus]`, else the first
+        // focusable control.
+        requestAnimationFrame(() => {
+            const editor = floater.querySelector('[autofocus]')
+                || floater.querySelector('input, textarea, select, button, [contenteditable=""], [contenteditable=true], [tabindex], komp-content-area, komp-select')
+            // preventScroll: the editor sits over the (already-visible) cell; focusing without
+            // it can scroll the grid to the editor's pre-positioned origin (the utilities layer).
+            editor?.focus?.({ preventScroll: true })
+            if (editor?.select && session.value == undefined) editor.select()
+            if (editor?.showPicker && editor.tagName === 'SELECT') { /* selects open on focus/click */ }
+
+        })
     }
-    
-    renderContextMenu (cell) {
-        if (cell.contextMenu) {
-            return cell.contextMenu(createElement('komp-spreadsheet-context-menu', {
-                content: [
-                    listenerElement('button', {
-                        name: 'copy',
-                        type: 'button',
-                        content: 'Copy',
-                        disabled: !cell.canCopy || !cell.canCopy()
-                    }, e => {
-                        this.copyCells()
-                        if (this.contextMenu) {
-                            this.contextMenu.remove()
-                            delete this.contextMenu
-                        }
-                    }),
-                    listenerElement('button', {
-                        name: 'paste',
-                        type: 'button',
-                        content: 'Paste',
-                        disabled: !cell.canPaste || !(cell.canPaste() && (window.navigator.clipboard.readText != undefined || this.copyData))
-                    }, async e => {
-                        window.navigator.permissions.query({
-                            name: 'clipboard-read',
-                            requestedOrigin: window.location.origin,
-                        }).then(async () => {
-                            this.pasteData(await window.navigator.clipboard.readText())
-                        }, () => {
-                            this.renderPasteModal();
+
+    /**
+     * End the active editing session — the single exit path for Enter/Tab, an editor's
+     * `session.commit()`/`session.cancel()`, and the `commit` event. Hides the editor
+     * floater (its `hidden` callback re-renders the cell once any held work settles and
+     * fires `cellChanged`), then moves the active cell or returns focus to the edited one.
+     *
+     * @param {Object} [options={}]
+     * @param {string} [options.move] - direction for {@link Spreadsheet#moveActive} ('up'/'down'/'left'/'right'); omitted, focus returns to the cell
+     */
+    deactivateCell({ move } = {}) {
+        if (!this._editing) return
+        const { cell, floater } = this._editing
+        floater.hide()
+        if (move) this.moveActive(move)
+        else cell.focus({ preventScroll: true })
+    }
+
+    /* -----------------------------------------------------------------
+       Copy / paste / clear (record-driven, work across off-screen ranges)
+    ----------------------------------------------------------------- */
+    quote(v) {
+        return (typeof v === 'string' && (v.includes('\n') || v.includes('\t'))) ? '"' + v + '"' : v
+    }
+
+    /**
+     * TSV for the selection. Resolves each column's record per row (so off-screen rows are
+     * included and resolver columns serialize their sub-record). Returns a Promise; it
+     * settles synchronously when records are already loaded.
+     */
+    async copyText() {
+        const r = this.range()
+        if (!r) return ''
+        const lines = []
+        for (let row = r.r0; row <= r.r1; row++) {
+            const cols = []
+            for (let col = r.c0; col <= r.c1; col++) {
+                const column = this.columns[col]
+                cols.push(this.quote(column.valueFor(await this.rows[row].resolvedRecordFor(column))))
+            }
+            lines.push(cols.join('\t'))
+        }
+        return lines.join('\n')
+    }
+
+    copyCells() {
+        if (!this.selection) return
+        // Show the marquee immediately; copyText is async (records may resolve), so write the
+        // clipboard once it settles — via the async Clipboard API (can't use the copy event's
+        // clipboardData after an await).
+        this._copying = true
+        this.paintSelection()
+        this.copyText().then(tsv => { if (navigator.clipboard) navigator.clipboard.writeText(tsv) })
+    }
+
+    pasteData(data) {
+        if (data == null) return
+        const r = this.range() || (this.active && { r0: this.active.row, r1: this.active.row, c0: this.active.col, c1: this.active.col })
+        if (!r) return
+        let matrix
+        if (typeof data == 'string') {
+            matrix = splitByUnquotedChar(data, "\n").map(r => splitByUnquotedChar(r, "\t"))
+        } else {
+            matrix = [[data]]
+        }
+
+        if (matrix.length > 1 && matrix[matrix.length - 1].length === 1 && matrix[matrix.length - 1][0] === '') {
+            matrix.pop()
+        }
+        const srcRows = matrix.length
+        const srcCols = Math.max(0, ...matrix.map(m => m.length))
+        if (!srcRows || !srcCols) return
+
+        // Tile the copied block in whole segments to fill the target range; any partial
+        // segment beyond the last whole repeat is left empty (e.g. a 1x3 block into a 1x7
+        // target fills two segments — [1,2,3,1,2,3] — leaving the 7th cell empty). A target
+        // smaller than the block still pastes one whole copy (extending past the selection).
+        const repRows = Math.max(1, Math.floor((r.r1 - r.r0 + 1) / srcRows))
+        const repCols = Math.max(1, Math.floor((r.c1 - r.c0 + 1) / srcCols))
+
+        for (let tr = 0; tr < repRows; tr++) {
+            for (let tc = 0; tc < repCols; tc++) {
+                this.pasteMatrixAt(matrix, r.r0 + tr * srcRows, r.c0 + tc * srcCols)
+            }
+        }
+
+        this.active = { row: r.r0, col: r.c0 }
+        this.cursor = {
+            row: Math.min(this.rows.length - 1, r.r0 + repRows * srcRows - 1),
+            col: Math.min(this.columns.length - 1, r.c0 + repCols * srcCols - 1)
+        }
+        this.setSelection(this.active, this.cursor) // selects the filled region; clears the copy marquee
+        this.trigger('cellChanged', { detail: { range: this.range() } })
+    }
+
+    pasteMatrixAt(matrix, row0, col0) {
+        matrix.forEach((cols, dr) => cols.forEach(async (data, dc) => {
+            const row = this.rows[row0 + dr]
+            const column = this.columns[col0 + dc]
+            if (!row || !column) return
+            if (this.isRecordReadonly(row.record)) return
+            const record = await row.resolvedRecordFor(column)
+            if (record == null) return
+            if (typeof data == 'string') {
+                if (column.accepts('text/plain')) {
+                    await column.paste(record, data)
+                } else {
+                    this.trigger('invalidPaste', { detail: data });
+                    return
+                }
+            } else if (data instanceof FileList) {
+                const files = []
+                for (const file of data) {
+                    if (column.accepts(file.type)) {
+                        files.push(file)
+                    } else {
+                        this.trigger('invalidPaste', {
+                            detail: data
                         })
-                    })
-                ]
-            }))
+                        return
+                    }
+                }
+                if (files.length > 0) {
+                    await column.paste(record, files)
+                }
+            }
+            
+            const cell = row.cellOf(column)
+            if (cell) cell.render()
+        }))
+    }
+
+    /**
+     * Read the system clipboard and paste it into the current selection. Use this for
+     * programmatic paste (e.g. a context-menu Paste action); the native `paste` event in
+     * {@link Spreadsheet#setupSelection} handles ⌘V/Ctrl+V directly. When the async
+     * Clipboard API or its permission is unavailable, fall back to {@link Spreadsheet#renderPasteModal}
+     * which explains the keyboard shortcuts.
+     */
+    async pasteFromClipboard() {
+        if (!window.navigator.permissions || !window.navigator.clipboard || !window.navigator.clipboard.readText) {
+            this.renderPasteModal()
+            return
+        }
+        try {
+            await window.navigator.permissions.query({
+                name: 'clipboard-read',
+                requestedOrigin: window.location.origin,
+            })
+            this.pasteData(await window.navigator.clipboard.readText())
+        } catch (e) {
+            this.renderPasteModal()
         }
     }
-    
-    renderPasteModal () {
+
+    /**
+     * Fallback shown when the clipboard can't be read programmatically — explains the
+     * ⌘C/⌘X/⌘V (or Ctrl) shortcuts the user can still use.
+     */
+    renderPasteModal() {
         const modifier = window.navigator.userAgent.toLowerCase().includes('mac') ? "&#8984;" : "Ctrl"
         this.getRootNode().body.append(new Modal({
             class: 'paste-modal',
@@ -379,170 +681,198 @@ export default class Spreadsheet extends Table {
             ]
         }))
     }
-    
-    /* -----------------
-        Copy / Paste
-    ----------------- */
-    async copyCells () {
-        const selectedCells = Array.from(this.selectedCells()).filter(c => c.canCopy())
-        const rows = Object.values(groupBy(selectedCells, 'rowIndex')).map(async row => {
-            return (await Promise.all(row.map(async cell => {
-                const v = await cell.copy()
-                if (typeof v == "string" && (v.includes("\n") || v.includes("\t"))) {
-                    return "\"" + v + "\""
-                }
-                return v
-            }))).join("\t")
-        })
-        const data = (await Promise.all(rows)).join("\n")
-        window.navigator.clipboard.writeText(data)
-        this.clearSelectedCells('')
-        const outline = this.outlineCells(selectedCells)
-        outline.classList.add('copy')
-    }
-    
-    pasteData(data) {
-        if (data == undefined) { return }
-        const cellMatrix = Object.values(groupBy(Array.from(this.selectedCells()).filter(c => c.canPaste()), 'rowIndex'))
-        let dataMatrix
-        if (typeof data == 'string') {
-            dataMatrix = splitByUnquotedChar(data, "\n").map(r => splitByUnquotedChar(r, "\t"))
-        } else {
-            dataMatrix = [[data]]
+
+    async clearSelectedCells() {
+        const r = this.range()
+        if (!r) return
+        for (let row = r.r0; row <= r.r1; row++) {
+            if (this.isRecordReadonly(this.rows[row].record)) continue
+            for (let col = r.c0; col <= r.c1; col++) {
+                const column = this.columns[col]
+                await column.clear(await this.rows[row].resolvedRecordFor(column))
+                const cell = this.rows[row].cellOf(column)
+                if (cell) cell.render()
+            }
         }
-        
-        cellMatrix.forEach((row, rowIndex) => {
-            const dataRowIndex = rowIndex % dataMatrix.length
-            row.forEach(async (cell, colIndex) => {
-                if (cell.paste) {
-                    const dataColIndex = colIndex % dataMatrix[dataRowIndex].length
-                    const data = dataMatrix[dataRowIndex][dataColIndex]
-                    if (typeof data == 'string') {
-                        if (cell.column.constructor.pasteAccepts.some(x => "text/plain".match(new RegExp(x)))) {
-                            await cell.paste(data)
-                        } else {
-                            this.trigger('invalidPaste', {
-                                detail: data
-                            })
-                        }
-                    } else if (data instanceof FileList) {
-                        const files = []
-                        for (const file of data) {
-                            if (cell.column.constructor.pasteAccepts.some(x => file.type.match(new RegExp(x)))) {
-                                files.push(file)
-                            } else {
-                                this.trigger('invalidPaste', {
-                                    detail: data
-                                })
-                            }
-                        }
-                        if (files.length > 0) {
-                            await cell.paste(files)
-                        }
-                    } else {
-                        return
-                    }
-                    cell.render()
-                }
-            })
-        })
-        this.querySelector(`${this.localName}-outline.copy`)?.remove()
+        this.trigger('cellChanged', { detail: { range: r } })
     }
 
-    static style = `
-        komp-spreadsheet {
+    /* -----------------------------------------------------------------
+       Context menu (right-click) — Copy / Paste, with a per-column hook
+    ----------------------------------------------------------------- */
+    activateContextMenu(e) {
+        const cell = e.delegateTarget
+        const menu = this.renderContextMenu(cell)
+        if (!menu) return
+        if (this.contextMenu) {
+            this.contextMenu.hide()
+            delete this.contextMenu
+        }
+        const targetBB = cell.getBoundingClientRect()
+        this.contextMenu = new Floater({
+            class: `${this.localName}-context-menu-floater`,
+            content: menu,
+            anchor: cell,
+            container: this.utilities, // grid-level layer above the header/frozen z-stack
+            offset: {
+                mainAxis: e.offsetX - targetBB.width,
+                crossAxis: e.offsetY
+            },
+            placement: 'right-start',
+            shift: false,
+            flip: true,
+            autoPlacement: false,
+            removeOnBlur: true,
+            autoUpdate: false
+        })
+        this.contextMenu.show()
+    }
+
+    renderContextMenu(cell) {
+        // Nothing selected → no menu.
+        if (!this.selection || !cell.column) return null
+        const menu = createElement(`${this.localName}-context-menu`, {
+            content: [
+                listenerElement('button', {
+                    name: 'copy',
+                    type: 'button',
+                    content: 'Copy'
+                }, () => {
+                    this.copyCells()
+                    if (this.contextMenu) {
+                        this.contextMenu.remove()
+                        delete this.contextMenu
+                    }
+                }),
+                listenerElement('button', {
+                    name: 'paste',
+                    type: 'button',
+                    content: 'Paste'
+                }, () => {
+                    // Reads the clipboard programmatically; falls back to the
+                    // keyboard-shortcut modal when that's unavailable (see #55).
+                    this.pasteFromClipboard()
+                    if (this.contextMenu) {
+                        this.contextMenu.remove()
+                        delete this.contextMenu
+                    }
+                })
+            ]
+        })
+        // Per-column customization hook (mirrors Spreadsheet): a column type may
+        // augment or replace the menu.
+        return cell.column.contextMenu(menu, cell, cell.column)
+    }
+
+    static style = function () { return `
+        ${this.tagName} {
             --select-oklch: 57.37% 0.1946 257.86;
             --select-color: oklch(var(--select-oklch));
-            --scroll-start: 0px;
-            overflow: auto;
-            scroll-behavior: smooth;
         }
-        komp-spreadsheet.scroll-snap {
-            scroll-snap-type: both mandatory;
-            scroll-padding-inline-start: var(--scroll-start);
-        }
-        komp-spreadsheet.scroll-snap komp-spreadsheet-cell {
-            scroll-snap-stop: always;
-            scroll-snap-align: start;
-        }
-        
-        komp-spreadsheet-header-cell,
-        komp-spreadsheet-cell {
+        ${this.tagName}-cell {
             cursor: cell;
             user-select: none;
             position: relative;
         }
-        komp-spreadsheet-focus {
-            outline: 2px solid var(--select-color);
-            outline-offset: -1px;
-            pointer-events: none !important;
-            grid-row: 1;
-        }
-        komp-spreadsheet-focus.frozen {
-            position: sticky;
-        }
-        komp-spreadsheet-focus.readonly {
-            outline-color: var(--disabled-color, #cecece)
-        }
-        
-        komp-spreadsheet-cell.selected::after,
-        komp-spreadsheet-cell.selecting::after {
+        ${this.tagName}-cell.selected::after {
             content: '';
             position: absolute;
             inset: 0;
-            background: oklch(var(--select-oklch) / 0.1);
-            z-index: 30;
             pointer-events: none;
+            background: color-mix(in oklab, var(--ds-select, #1a73e8) 12%, transparent);
         }
-        komp-spreadsheet komp-table-header > *:focus,
-        komp-spreadsheet komp-table-header > *.selecting,
-        komp-spreadsheet komp-table-header > *.selected {
-            box-shadow: inset 0 2px 0 0 var(--select-color), inset 0 0 0 999px oklch(var(--select-oklch) / 0.1);
-            outline: none !important;
+        ${this.tagName}-cell.active {
+            outline: 2px solid var(--select-color);
+            outline-offset: -1px;
+            z-index: 3;
         }
-        
-        .komp-spreadsheet-input {
-            position: relative;
-        }
-        .komp-spreadsheet-input > label {
-            min-height: var(--cell-height);
-            min-width: var(--cell-width);
-            width: auto;
-            display: flex;
-            flex-direction: column;
-            justify-content: space-between;
-            background: white;
-            box-shadow: 0 0 0 3px oklch(var(--select-oklch) / 0.35);
+        /* The focused cell is outlined. moveActive keeps focus on the active cell while
+           extending (so this coincides with .active — one outline, not two). */
+        ${this.tagName}-cell:focus {
             outline: 2px solid var(--select-color);
             outline-offset: -1px;
         }
-        .komp-spreadsheet-input > label > input,
-        .komp-spreadsheet-input > label > textarea,
-        .komp-spreadsheet-input > label > select,
-        .komp-spreadsheet-input > label > komp-content-area,
-        .komp-spreadsheet-input > label > label {
-            background: none;
-            width: auto;
-            min-height: 100%;
-            min-width: 100%;
-            outline: none;
-            padding: var(--padding, unset);
-            border: none;
+        ${this.tagName}-cell.readonly {
+            cursor: default;
         }
-        .komp-spreadsheet-input komp-content-area {
-            width: max-content;
-            max-width: var(--remaining-width);
-            max-height: var(--remaining-height);
+        ${this.tagName}-cell.readonly.active,
+        ${this.tagName}-cell.readonly:focus {
+            outline-color: var(--disabled-color, #cecece);
         }
-        .komp-spreadsheet-input > label > input {
-            width: 1px; // make so doesn't initially overflow column
-        }
-        
-        komp-spreadsheet .number-cell {
+        ${this.tagName}-cell.number-cell {
             text-align: right;
         }
+        ${this.tagName}-cell.checkbox-cell {
+            text-align: center;
+        }
 
-        komp-spreadsheet-context-menu {
+        /* Border box encompassing the selected range (positioned in body space by
+           renderSelectionIndicator, so it spans off-screen cells). */
+        ${this.tagName}-select-indicator {
+            position: absolute;
+            pointer-events: none;
+            box-sizing: border-box;
+            outline: 1px solid var(--select-color);
+            z-index: 5;
+        }
+        ${this.tagName}-select-indicator.copy {
+            outline-style: dashed;
+            outline-width: 2px;
+            box-shadow: 0 0 0 1px var(--select-color);
+        }
+
+        ${this.tagName}-row:has(${this.tagName}-cell.active) {
+            z-index: 4;
+        }
+
+        /* Editor floater: a bordered box over the cell that can overflow the cell
+           bounds for larger input contexts (mirrors Spreadsheet's komp-spreadsheet-input). */
+        .${this.tagName}-input {
+            z-index: 4;
+        }
+        .${this.tagName}-input > label {
+            display: flex;
+            min-width: var(--cell-width, auto);
+            min-height: var(--cell-height, auto);
+            background: white;
+            outline: 2px solid var(--select-color);
+            outline-offset: -1px;
+            box-shadow: 0 0 0 3px oklch(var(--select-oklch) / 0.35);
+            
+            display: flex;
+            flex-direction: column;
+            justify-content: space-between;
+        }
+        .${this.tagName}-input > label > input,
+        .${this.tagName}-input > label > textarea,
+        .${this.tagName}-input > label > select,
+        .${this.tagName}-input > label > komp-content-area,
+        .${this.tagName}-input > label > komp-select > button {
+            font: inherit;
+            border: none;
+            outline: none;
+            background: none;
+            min-width: 100%;
+            min-height: 100%;
+            box-sizing: border-box;
+            padding: 0.4em 0.6em;
+            border-radius: unset;
+        }
+        .${this.tagName}-input > label > input {
+            width: 1px; // make so doesn't initially overflow column
+        }
+
+        komp-modal.paste-modal {
+            background: white;
+            padding: 1em;
+        }
+
+        /* Right-click context menu — a small floating button list, above the
+           selection/editor layers (mirrors komp-spreadsheet-context-menu). */
+        .${this.tagName}-context-menu-floater {
+            z-index: 6;
+        }
+        ${this.tagName}-context-menu {
             display: block;
             border-radius: 0.35em;
             background: white;
@@ -550,66 +880,31 @@ export default class Spreadsheet extends Table {
             font-size: 0.8em;
             box-shadow: 0 2px 12px 2px rgba(0,0,0, 0.2), 0 1px 2px 1px rgba(0,0,0, 0.3);
         }
-        komp-spreadsheet-context-menu > button {
+        ${this.tagName}-context-menu > button {
             display: block;
             width: 100%;
             outline: none;
             appearance: none;
             border: none;
-            background:none;
+            background: none;
             padding: 0.2em 0.5em;
             border-radius: 0.25em;
         }
-        komp-spreadsheet-context-menu > button:disabled {
+        ${this.tagName}-context-menu > button:disabled {
             opacity: 0.5;
         }
-        komp-spreadsheet-context-menu > button:disabled:hover {
+        ${this.tagName}-context-menu > button:disabled:hover {
             background: white;
             color: inherit;
         }
-        komp-spreadsheet-context-menu > button:focus,
-        komp-spreadsheet-context-menu > button:hover {
+        ${this.tagName}-context-menu > button:focus,
+        ${this.tagName}-context-menu > button:hover {
             background: oklch(var(--select-oklch) / 0.2);
         }
-        komp-spreadsheet-context-menu > button:hover {
+        ${this.tagName}-context-menu > button:hover {
             color: var(--select-color);
         }
-        
-        komp-spreadsheet-outline {
-            position: absolute;
-            pointer-events: none;
-            outline: 1px solid var(--select-color);
-        }
-        komp-spreadsheet-outline.copy {
-            outline: 2px dashed var(--select-color);
-        }
-        
-        komp-spreadsheet komp-floater {
-            user-select: text;
-        }
-        
-        
-        komp-spreadsheet-cell               { z-index: 1; }
-        komp-spreadsheet-focus              { z-index: 25; }
-        komp-spreadsheet-input              { z-index: 26; }
-
-        komp-spreadsheet-cell.frozen                   { z-index: 100; }
-        komp-spreadsheet-focus.frozen                  { z-index: 101; }
-        komp-spreadsheet-input.frozen                  { z-index: 102; }
-        komp-spreadsheet komp-table-header             { z-index: 110; }
-        komp-spreadsheet komp-table-header > .frozen   { z-index: 111; }
-        
-        komp-spreadsheet-outline            { z-index: 201; }
-        komp-spreadsheet komp-floater       { z-index: 300; }
-        komp-spreadsheet komp-tooltip       { z-index: 300; }
-        
-        komp-modal.paste-modal {
-            background: white;
-            padding: 1em;
-        }
-    `
+    `}
 
     static { this.define() }
 }
-Spreadsheet.include(resizable)
-Spreadsheet.include(reorderable)
