+++ lib/komps/spreadsheet/column.js
@@ -1,157 +1,101 @@
 /**
- * The configuration class for {@link Spreadsheet} columns
+ * Column controller for a {@link Spreadsheet} — extends {@link DataGridColumn}
+ * with a **record-based editing API**.
+ *
+ * In a virtualized grid the cell for a given record may not be mounted (off-screen),
+ * so editing/paste operate on the **record**, not a live cell. Otherwise it mirrors
+ * Spreadsheet's column shape: `input(record)` builds a bound editor via {@link Input},
+ * `paste`/`parsePaste` apply pasted values, `clear` empties.
  *
  * @class SpreadsheetColumn
- * @extends TableColumn
+ * @extends DataGridColumn
  *
  * @param {Object} [options={}]
- * @param {boolean} [options.headerEditable=true] - allow header to be changed, fires events when changed
- *
- * @example <caption>Custom Column Type</caption>
- * Spreadsheet.columnTypeRegistry.foo = class FooColumn extends SpreadsheetColumn {
- *     ...
- * }
- * new Spreadsheet({
- *     data: [...]
- *     columns: [{ type: 'foo' }]
- * })
+ * @param {string}  [options.attribute] - record key this column reads/writes
+ * @param {boolean} [options.editable=true] - whether cells can be edited/pasted into
+ * @param {Object}  [options.input] - extra options merged into Input.create (e.g. `dump`/`load`)
+ * @param {Function} [options.contextMenu] - hook `(menu, cell, column)` to augment/replace the right-click menu
  */
 
-import SpreadsheetCell from './cell.js';
-import HeaderCell from './header-cell.js';
-import TableColumn from '../table/column.js';
-import Input from '../input.js';
-import { result } from '../../support.js';
+import DataGridColumn from '../data-grid/column.js'
+import Input from '../input.js'
 
-let onEnterDeprecationWarned = false;
-
-export default class SpreadsheetColumn extends TableColumn {
-    static Cell = SpreadsheetCell;
-    static HeaderCell = HeaderCell;
-    
-    static assignableAttributes = {
-        type: { type: 'string', default: null, null: true },
-        attribute: { type: 'string', default: null, null: true },
-        headerEditable: { type: 'boolean', default: true, null: false }
-    }
-    
-    static assignableMethods = [
-        'input', 'copy', 'paste'
-    ]
-    
+export default class SpreadsheetColumn extends DataGridColumn {
+    /** Input kind passed to `Input.create`; overflow-friendly rich text by default (mirrors Spreadsheet). */
+    static inputType = 'contentarea'
+    /** Extra `Input.create` options merged into every editor (override per type, e.g. number `dump`). */
     static inputOptions = {}
-    
-    initialize (configs) {
-        this.configuredContextMenu = configs.contextMenu
-        this.inputOptions = Object.assign({}, configs.input, this.constructor.inputOptions)
-        if (configs.hasOwnProperty('copy') && !configs.copy) {
-            this.copy = undefined
-        }
-        if (configs.hasOwnProperty('paste') &&!configs.paste) {
-            this.paste = undefined
-        }
+
+    constructor(options = {}) {
+        super(options)
+        if (this.editable === undefined) this.editable = true
+        this.inputOptions = Object.assign({}, this.constructor.inputOptions, options.input)
+        this.configuredContextMenu = options.contextMenu
     }
 
-    headerFallback () {
-        return this.attribute
-    }
-    
-    render (record) {
-        const value = result(record, this.attribute)
-        if (value && typeof value == 'string') {
-            return value.replaceAll("\n", "<br>")
+    /**
+     * Per-column customization of the right-click context menu (mirrors Spreadsheet).
+     * Receives the built menu element and returns the (possibly augmented or replaced)
+     * menu. Defaults to returning the menu unchanged; supply a `contextMenu` option to
+     * override.
+     *
+     * @param {HTMLElement} menu
+     * @param {SpreadsheetCell} cell
+     * @param {SpreadsheetColumn} column
+     * @returns {HTMLElement}
+     */
+    contextMenu(menu, ...args) {
+        if (this.configuredContextMenu) {
+            return this.configuredContextMenu(menu, ...args)
         }
-        return value
+        return menu
     }
-    
-    copy (cell) {
-        return result(cell.record, this.attribute)
+
+    /** Parse a pasted/typed string into the stored value. Override per type. */
+    parsePaste(value) { return value }
+
+    /** Apply a pasted value to the record. */
+    paste(record, value) {
+        if (this.attribute != null) record[this.attribute] = this.parsePaste(value)
     }
-    
-    static pasteAccepts = ["text/plain"]
-    paste (cell, value) {
-        cell.record[this.attribute] = value
+
+    /** Clear the record's value. */
+    clear(record) {
+        if (this.attribute != null) record[this.attribute] = null
     }
-    
-    clear (cell) {
-        cell.record[this.attribute] = null
+
+    /** MIME types this column accepts on paste. */
+    static pasteAccepts = ['text/plain']
+    accepts(type = 'text/plain') {
+        return this.editable && this.constructor.pasteAccepts.some(p => type.match(new RegExp(p)))
     }
-    // TODO move to cell based signature (cell, options)
-    input (record, cell, options) {
-        return Input.create(this.type || 'contentarea', Object.assign({
+
+    /**
+     * Build the bound edit control for a record (mirrors Spreadsheet via `Input.create`).
+     * The returned element is two-way bound to `record[attribute]` — edits write back on
+     * change/blur — and is shown in a {@link Floater} over the cell. Returns `null` when
+     * the column isn't editable.
+     *
+     * Custom editors can take over completion via the session (see
+     * {@link Spreadsheet#activateCell}):
+     *  - keep a key for yourself (e.g. Enter for newlines in a textarea) by calling
+     *    `stopPropagation()` or `preventDefault()` on its `keydown`
+     *  - call `session.commit({ move })` / `session.cancel()` to end the edit, or
+     *    `session.hold(promise)` to defer the closing re-render until async work settles
+     *
+     * Pick what you need off the session — don't spread it into Input/element options,
+     * where its members would become HTML attributes on the editor.
+     *
+     * @param {Object} record
+     * @param {Object} [session] - editing session: `{ cell, column, record, floater, value, commit, cancel, hold }` (`value` is the typed character seeding a type-to-edit)
+     * @returns {HTMLElement|null}
+     */
+    input(record, options = {}) {
+        if (this.editable === false) return null
+        return Input.create(this.constructor.inputType, Object.assign({
             target: record,
             attribute: this.attribute,
             autofocus: true
-        }, this.inputOptions, options))
-    }
-    /**
-     * Called after a cell spawns its input floater and focus has been set up.
-     * Receives the cell and its floater directly, so column types can wire
-     * input-specific keyboard and lifecycle behavior without reaching through
-     * the DOM. Override to customize; not calling `super` opts out of the
-     * default navigation below.
-     *
-     * Default behavior provides spreadsheet keyboard navigation: Escape cancels
-     * the edit and refocuses the cell, Enter commits and moves to the cell below.
-     *
-     * @param {SpreadsheetCell} cell
-     * @param {Floater} floater
-     */
-    cellActivated (cell, floater) {
-        const table = cell.table
-        floater.addEventListener('keyup', e => {
-            if (e.key == "Escape") {
-                floater.setAttribute('preventChange', true)
-                cell.focus()
-                e.preventDefault()
-            } else if (
-                e.key == "Enter" &&
-                this._enterAllowed(e) &&
-                !table._enterDown &&
-                [e.metaKey,e.ctrlKey,e.altKey,e.shiftKey].every(x => x == false)
-            ) {
-                e.preventDefault()
-                const newCell = cell.focusAdjacentCell('down')
-                if (!newCell) {
-                    cell.focus()
-                    floater.hide()
-                }
-            }
-        })
-        floater.addEventListener('keydown', e => {
-            if (
-                e.key == "Enter" &&
-                this._enterAllowed(e) &&
-                [e.metaKey,e.ctrlKey,e.altKey,e.shiftKey].every(x => x == false)
-            ) {
-                e.preventDefault()
-            }
-        })
-    }
-
-    /**
-     * @deprecated Override {@link SpreadsheetColumn#cellActivated} instead.
-     * Returning `false` to suppress the default Enter navigation is still
-     * honored for backward compatibility but will be removed in a future release.
-     */
-    onEnter () {}
-
-    _enterAllowed (e) {
-        if (this.onEnter !== SpreadsheetColumn.prototype.onEnter) {
-            if (!onEnterDeprecationWarned) {
-                onEnterDeprecationWarned = true
-                console.warn(`[Komps] ${this.constructor.name}#onEnter is deprecated; override cellActivated(cell, floater) instead.`)
-            }
-            return this.onEnter(e) !== false
-        }
-        return true
-    }
-    
-    contextMenu (menu, ...args) {
-        if (this.configuredContextMenu) {
-            return this.configuredContextMenu(menu, ...args)
-        } else {
-            return menu
-        }
+        }, this.inputOptions, options)) // pass all options so using app can pass what it wants through
     }
 }
