+++ lib/komps/spreadsheet/columns/number-column.js
@@ -1,34 +1,27 @@
-import Column from '../column.js';
-import { result } from '../../../support.js';
+/**
+ * Numeric column for {@link Spreadsheet}. Edits with a numeric input that dumps
+ * to a float; parses pasted values as floats; renders right-aligned via `number-cell`.
+ *
+ * @class NumberColumn
+ * @extends SpreadsheetColumn
+ */
+import SpreadsheetColumn from '../column.js'
 
-export default class NumberColumn extends Column {
+function toNumber(v) {
+    if (v == null || v === '') return null
+    if (typeof v === 'number') return v
+    const n = String(v).includes('.') ? parseFloat(v) : parseInt(v, 10)
+    return Number.isNaN(n) ? null : n
+}
 
-    static assignableAttributes = {
-        type: { type: 'string', default: 'number', null: false }
-    }
-    
-    static inputOptions = {
-        dump: v => {
-            if (typeof v == 'string') {
-                return v.includes(".") ? parseFloat(v) : parseInt(v)
-            }
-            return v
-        }
-    }
+export default class NumberColumn extends SpreadsheetColumn {
+    static inputType = 'number'
+    static inputOptions = { dump: toNumber }
 
-    render (record) {
-        const value = result(record, this.attribute)
-        if (value != null && typeof value == 'number') {
-            return value.toLocaleString()
-        }
-        return value
+    constructor(options = {}) {
+        super(options)
+        this.class = [this.class, 'number-cell'].filter(Boolean).join(' ')
     }
 
-    async paste (cell, value) {
-        const v = parseFloat(value)
-        if (!isNaN(v)) {
-            return super.paste(cell, v)
-        }
-        return super.paste(cell, value)
-    }
+    parsePaste(value) { return toNumber(value) }
 }
