+++ lib/komps/spreadsheet/columns/readonly-column.js
@@ -1,10 +1,19 @@
-import TableColumn from '../column.js';
+/**
+ * Non-editable column for {@link Spreadsheet}: cells display but can't be
+ * activated, pasted into, or cleared.
+ *
+ * @class ReadonlyColumn
+ * @extends SpreadsheetColumn
+ */
+import SpreadsheetColumn from '../column.js'
 
-export default class ReadonlyColumn extends TableColumn {
-    paste = undefined;
-    
-    renderCell (record, options={}) {
-        options.readonly = true;
-        return super.renderCell(record, options)
+export default class ReadonlyColumn extends SpreadsheetColumn {
+    constructor(options = {}) {
+        super(Object.assign({ editable: false }, options))
     }
+
+    input() { return null }
+    paste() {}
+    clear() {}
+    accepts() { return false }
 }
