+++ lib/komps/spreadsheet/columns/select-column.js
@@ -1,27 +1,36 @@
-import Column from '../column.js';
+/**
+ * Dropdown column for {@link Spreadsheet}. Edits with a bound `<select>`
+ * (Input.create) built from `options`.
+ *
+ * @class SelectColumn
+ * @extends SpreadsheetColumn
+ *
+ * @param {Object} [options={}]
+ * @param {Array} [options.options=[]] - allowed values; strings, or `{ value, label }`
+ */
+import SpreadsheetColumn from '../column.js'
 import Select from '../../select.js';
 
-export default class SelectColumn extends Column {
+export default class SelectColumn extends SpreadsheetColumn {
+    static inputType = 'select'
 
-    static inputAttributes = ['options']
-
-    static assignableAttributes = {
-        type: { type: 'string', default: 'select', null: false }
-    }
-
-    constructor (options) {
+    constructor(options = {}) {
         super(options)
-        this.options = options.options
+        this.options = options.options || []
     }
 
-    input (record, cell, options) {
-        const {value, ...scrubbedOptions} = options 
+    optionList() {
+        return this.options.map(o => (o != null && typeof o === 'object') ? o : { value: o, label: o })
+    }
+
+    /** Build a {@link Select} editor; `session.value` (type-to-edit) seeds the typeahead. */
+    input (record, { value } = {}) {
         const select = new Select(Object.assign({
             autofocus: true,
             target: record,
             attribute: this.attribute,
             options: this.options
-        }, this.inputOptions, scrubbedOptions))
+        }, this.inputOptions))
         if (value) {
             select.typeahead(value)
         }
@@ -29,34 +38,9 @@ export default class SelectColumn extends Column {
         return select
     }
 
-    // Diverges from the default Enter navigation: a dropdown option is chosen
-    // via its native Enter->click, and selecting it already closes the floater,
-    // so this only wires the close-on-change and Escape-to-cancel behavior.
-    cellActivated (cell, floater) {
-        const select = floater.querySelector('komp-select')
-        select.addEventListener('change', () => {
-            select.dropdown.hide()
-            floater.hide()
-            cell.focus()
-        })
-        floater.addEventListener('keyup', e => {
-            if (e.key == "Escape") {
-                floater.setAttribute('preventChange', true)
-                cell.focus()
-                e.preventDefault()
-            }
-        })
+    parsePaste(value) {
+        // accept a label or a value; map back to a known value when possible
+        const match = this.optionList().find(o => o.value === value || o.label === value)
+        return match ? match.value : value
     }
-    
-    static style = `
-        .komp-spreadsheet-input > label > komp-select > button {
-            background: none;
-            width: auto;
-            min-height: 100%;
-            min-width: 100%;
-            outline: none;
-            padding: var(--padding, unset);
-            border: none;
-        }
-    `
 }
