--- lib/komps/spreadsheet/index.js
@@ -1,3 +1,0 @@
-export { default as Column } from './column.js'
-export { default as Cell } from './cell.js'
-export { default as HeaderCell } from './header-cell.js'
