--- lib/komps/table/header.js
@@ -1,16 +1,0 @@
-import TableRow from './row.js';
-
-export default class TableHeader extends TableRow {
-    static tagName = 'komp-table-header'
-
-    connected () {
-        super.connected();
-        this.setAttribute('role', 'row');
-    }
-
-    rowIndexChanged(was, now) {
-        this.style.gridRow = now
-    }
-
-    static { this.define() }
-}
