--- lib/komps/table/plugins/cellsDimensions.js
@@ -1,10 +1,0 @@
-export default function (proto) {
-    proto.cellsDimensions = function (dimension) {
-        const cells = this.querySelectorAll(this.cellSelector)
-        const lastCell = cells.item(cells.length - 1)
-        return {
-            width: lastCell.offsetLeft + lastCell.offsetWidth,
-            height: lastCell.offsetTop + lastCell.offsetHeight
-        }[dimension]
-    }
-}
